requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==3.2.25
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
Faker==29.0.0